+++ SideMolly/src-tauri/src/bundles.rs
@@ -1601,6 +1601,7 @@ mod tests {
         conn.execute_batch(include_str!("../migrations/009_bundle_file_rotation.sql")).unwrap();
         conn.execute_batch(include_str!("../migrations/010_jobs_kind_widen.sql")).unwrap();
         conn.execute_batch(include_str!("../migrations/011_auto_assembly_settings.sql")).unwrap();
+        conn.execute_batch(include_str!("../migrations/012_processing_log.sql")).unwrap();
         conn
     }